services/hr_assistant/requirements.txt
fastapi
uvicorn

fpdf
pandas
gspread 
google-auth
python-dotenv
oauth2client

openai
requests
seaborn